Oki: first implementation of bit feeder
It is really complicated stuff. General idea is to use masks (called
fill) inside of word to mark if byte is ready to be printed or no.

Then after each bit it is check in a loop if next byte in order could be
pushed to output. This loop works until actually set bit. If it find at
least one word that is not ready, it gives up.