Fix bug in binary reader
It was forgetting last few bits all the time. This is because it has
internally two counter: one for number of bits not read from big stream
into internal states and the other to track internal buffer, which is
always 0-8 bits.

The problem was that if whole buffer was read into internal states,
isnext() returned that the is no more bits in buffer. Now it checks also
internal buffers