Introduce descriptions
It is highly inconvenient to type parameter like 'Oki 3321' via command
line, as it have to be put in quotes. So we introduce description, where
user friendly name is presented and in addition to than name will be
used to set more friendly names, i.e. with no uppercase and spaces.